Some small changes

+++ code/handmade.cpp
@@ -12,8 +12,8 @@ RenderWeirdGradient(game_offscreen_buffer *Buffer, int BlueOffset, int GreenOffs
         
         for (int X = 0; X < Buffer->Width; X++)
         {
-            uint8 Blue = (X + BlueOffset);
-            uint8 Green = (Y + GreenOffset);
+            uint8 Blue = (uint8)(X + BlueOffset);
+            uint8 Green = (uint8)(Y + GreenOffset);
             
             *Pixel++ = ((Green << 8) | Blue);
         }
@@ -65,7 +65,7 @@ GameUpdateAndRender(game_memory *Memory, game_input *Input, game_offscreen_buffe
     game_controller_input *Input0 = &Input->Controllers[0];
     if(Input0->IsAnalog)
     {
-        GameState->BlueOffset += (int)4.0f*(Input0->EndX);
+        GameState->BlueOffset += (int)(4.0f*(Input0->EndX));
     }
     else
     {
